+++ dlight.dtrace/src/org/netbeans/modules/dlight/dtrace/collector/support/DtraceDataAndStackParser.java
@@ -175,11 +175,11 @@ final class DtraceDataAndStackParser extends DtraceParser {
                     return null;
                 } else {
                     Collections.reverse(currStack);
-                    int stackId = sds == null? -1 : sds.putStack(currStack, currSampleDuration);
+                    long stackId = sds == null? -1 : sds.putStack(currStack, currSampleDuration);
                     currStack.clear();
                     //colNames.get(colNames.size()-1);
                     state = State.WAITING_DATA;
-                    currData.add(Integer.toString(stackId));
+                    currData.add(Long.toString(stackId));
                     return new DataRow(colNames, currData);
                 }
         }
